gb/mp-allowed instructors to add and update tutors on tutor assignment

+++ src/main/java/edu/ucsb/cs56/ucsb_open_lab_scheduler/controllers/TutorAssignmentController.java
@@ -66,12 +66,12 @@ public class TutorAssignmentController {
     String role = authControllerAdvice.getRole(token);
     String email= (String) token.getPrincipal().getAttributes().get("email");
     List<CourseOffering> instructorCourses=courseOfferingRepository.findByInstructorEmail(email);
-    boolean n=false;
+    boolean isInstructorForCourse=false;
     for(int i=0;i<instructorCourses.size();i++){
       if(instructorCourses.get(i).getId()==id)
-        n=true;
+        isInstructorForCourse=true;
     }
-    if (!((role.equals("Admin")) || (n))) {
+    if (!((role.equals("Admin")) || (isInstructorForCourse))) {
       redirAttrs.addFlashAttribute("alertDanger", "You do not have permission to access that page");
       return "redirect:/";
     }
@@ -108,7 +108,15 @@ public class TutorAssignmentController {
   public ResponseEntity<?> add(@RequestParam("cid") long cid, @RequestParam("tid") long tid,  @RequestParam("lead") boolean lead,
                                OAuth2AuthenticationToken token) {
     String role = authControllerAdvice.getRole(token);
-    if (!(role.equals("Admin"))) {
+    String email= (String) token.getPrincipal().getAttributes().get("email");
+    List<CourseOffering> instructorCourses=courseOfferingRepository.findByInstructorEmail(email);
+    boolean isInstructorForCourse=false;
+    for(int i=0;i<instructorCourses.size();i++){
+      if(instructorCourses.get(i).getId()==cid)
+        isInstructorForCourse=true;
+    }
+
+    if (!((role.equals("Admin"))||(isInstructorForCourse))) {
         return new ResponseEntity<>(HttpStatus.FORBIDDEN);
     }
 
@@ -127,9 +135,19 @@ public class TutorAssignmentController {
     public ResponseEntity<?> update(@RequestParam("cid") long cid, @RequestParam("tid") long tid, @RequestParam("lead") boolean lead,
                                  OAuth2AuthenticationToken token) {
       String role = authControllerAdvice.getRole(token);
-      if (!role.equals("Admin")) {
-          return new ResponseEntity<>(HttpStatus.FORBIDDEN);
-      }
+    String email= (String) token.getPrincipal().getAttributes().get("email");
+    List<CourseOffering> instructorCourses=courseOfferingRepository.findByInstructorEmail(email);
+    boolean isInstructorForCourse=false;
+    for(int i=0;i<instructorCourses.size();i++){
+      if(instructorCourses.get(i).getId()==cid)
+        isInstructorForCourse=true;
+    }
+
+    if (!((role.equals("Admin"))||(isInstructorForCourse))) {
+        return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+    }
+
+
 
       Tutor tutor = tutorRepository.findById(tid)
           .orElseThrow(() -> new IllegalArgumentException("Invalid tutor Id:" + tid));
@@ -148,8 +166,16 @@ public class TutorAssignmentController {
   public ResponseEntity<?> delete(@PathVariable("cid") long cid, @PathVariable("tid") long tid,
   OAuth2AuthenticationToken token) {
     String role = authControllerAdvice.getRole(token);
-    if (!(role.equals("Admin"))) {
-      return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+    String email= (String) token.getPrincipal().getAttributes().get("email");
+    List<CourseOffering> instructorCourses=courseOfferingRepository.findByInstructorEmail(email);
+    boolean isInstructorForCourse=false;
+    for(int i=0;i<instructorCourses.size();i++){
+      if(instructorCourses.get(i).getId()==cid)
+        isInstructorForCourse=true;
+    }
+
+    if (!((role.equals("Admin"))||(isInstructorForCourse))) {
+        return new ResponseEntity<>(HttpStatus.FORBIDDEN);
     }
 
     tutorAssignmentRepository.deleteByCourseOfferingIdAndTutorId(cid, tid);
